refactor(scheduler): separate actor terminal and wake policies
