+++ rust-srec/src/scheduler/actor/streamer_actor.rs
@@ -37,6 +37,14 @@ use crate::monitor::{LiveStatus, ProcessStatusResult, ProcessStatusSuppression};
 use crate::scheduler::actor::DownloadEndPolicy;
 use crate::streamer::StreamerMetadata;
 
+mod batch;
+mod checks;
+mod feedback;
+mod mailbox;
+mod run;
+mod terminal;
+mod wake;
+
 /// Result type for actor operations.
 pub type ActorResult = Result<ActorOutcome, ActorError>;
 
@@ -293,2812 +301,7 @@ impl StreamerActor {
             .map(|m| m.consecutive_error_count as u32)
             .unwrap_or(0)
     }
-
-    /// Run the actor's event loop.
-    ///
-    /// This method runs until the actor receives a Stop message or the
-    /// cancellation token is triggered.
-    ///
-    /// # Returns
-    ///
-    /// Returns `ActorOutcome::Stopped` on graceful shutdown,
-    /// `ActorOutcome::Cancelled` if cancelled externally.
-    pub async fn run(mut self) -> ActorResult {
-        info!("StreamerActor {} starting", self.id);
-
-        // Schedule initial check if not already scheduled
-        if self.state.next_check.is_none() {
-            self.state
-                .schedule_next_check(&self.config, self.get_error_count());
-        }
-
-        loop {
-            // First, drain all priority messages before processing normal messages
-            // This ensures high-priority operations (like Stop) are handled promptly
-            if let Some(msg) = self.try_recv_priority() {
-                let start = Instant::now();
-                let should_stop = self.handle_run_message(msg).await?;
-                self.metrics.record_message(start.elapsed());
-
-                if should_stop {
-                    debug!(
-                        "StreamerActor {} received stop signal from priority channel",
-                        self.id
-                    );
-                    break;
-                }
-                // Continue to check for more priority messages
-                continue;
-            }
-
-            // Calculate sleep duration before entering select to avoid borrow issues.
-            // Returns None if no check is scheduled (e.g., when streamer is live).
-            //
-            // To avoid getting stuck indefinitely in Live state when external download
-            // orchestration fails to send DownloadEnded, perform an occasional watchdog
-            // check while Live. Also, if the download appears stalled (no heartbeat),
-            // check sooner to recover state quickly.
-            let mut sleep_duration = self.state.time_until_next_check();
-            if sleep_duration.is_none() && self.state.streamer_state == StreamerState::Live {
-                let watchdog = self.live_watchdog_interval();
-                let stall = self.time_until_live_stall_watchdog();
-                let mut next = std::cmp::min(watchdog, stall);
-
-                // Boundary wake: if the last status check provided a hint (e.g. schedule end
-                // time), wake up at that time even while Live.
-                if let Some(ref last) = self.state.last_check
-                    && let Some(hint) = last.next_check_hint
-                {
-                    let now = chrono::Utc::now();
-                    if hint <= now {
-                        next = Duration::ZERO;
-                    } else if let Ok(delay) = hint.signed_duration_since(now).to_std() {
-                        next = std::cmp::min(next, delay);
-                    }
-                }
-
-                // Hold off the watchdog after a failed check: overrides the stall/hint
-                // ZERO wake so an unreachable download does not drive a tight retry loop.
-                if let Some(backoff_until) = self.live_watchdog_backoff_until {
-                    next = std::cmp::max(
-                        next,
-                        backoff_until.saturating_duration_since(Instant::now()),
-                    );
-                }
-
-                sleep_duration = Some(next);
-            }
-            let check_timer = Self::create_check_timer(sleep_duration);
-
-            tokio::select! {
-                // Bias towards handling messages first
-                biased;
-
-                // Check priority mailbox first (if configured)
-                Some(msg) = Self::recv_priority_opt(&mut self.priority_mailbox) => {
-                    let start = Instant::now();
-                    let should_stop = self.handle_run_message(msg).await?;
-                    self.metrics.record_message(start.elapsed());
-
-                    if should_stop {
-                        debug!("StreamerActor {} received stop signal from priority channel", self.id);
-                        break;
-                    }
-                }
-
-                // Handle normal-priority messages
-                Some(msg) = self.mailbox.recv() => {
-                    let start = Instant::now();
-                    let should_stop = self.handle_run_message(msg).await?;
-                    self.metrics.record_message(start.elapsed());
-
-                    if should_stop {
-                        debug!("StreamerActor {} received stop signal", self.id);
-                        break;
-                    }
-                }
-
-                // Self-scheduled check timer
-                _ = check_timer => {
-                    trace!(streamer_id = %self.id, "check timer fired");
-                    if let Err(e) = self.initiate_check().await {
-                        warn!("StreamerActor {} check failed: {}", self.id, e);
-                        self.metrics.record_error();
-
-                        // Fatal (non-recoverable) errors should stop the actor - the scheduler
-                        // will receive a StreamerStateSyncedFromDb event and clean up.
-                        if !e.recoverable {
-                            info!(
-                                "StreamerActor {} stopping due to fatal error: {}",
-                                self.id, e.message
-                            );
-                            break;
-                        }
-                    }
-                }
-
-                // Cancellation
-                _ = self.cancellation_token.cancelled() => {
-                    info!("StreamerActor {} cancelled", self.id);
-                    return Ok(ActorOutcome::Cancelled);
-                }
-            }
-        }
-
-        info!("StreamerActor {} stopped gracefully", self.id);
-        Ok(ActorOutcome::Stopped)
-    }
-
-    /// Fatal mailbox errors are terminal policy decisions, just like fatal timer errors.
-    /// Return a stop signal so all mailbox paths reach graceful shutdown.
-    async fn handle_run_message(&mut self, message: StreamerMessage) -> Result<bool, ActorError> {
-        match self.handle_message(message).await {
-            Err(error) if !error.recoverable => {
-                self.metrics.record_error();
-                info!(streamer_id = %self.id, error = %error, "Stopping actor after terminal message error");
-                Ok(true)
-            }
-            result => result,
-        }
-    }
-
-    /// Try to receive a message from the priority mailbox without blocking.
-    fn try_recv_priority(&mut self) -> Option<StreamerMessage> {
-        if let Some(ref mut priority_rx) = self.priority_mailbox {
-            priority_rx.try_recv().ok()
-        } else {
-            None
-        }
-    }
-
-    /// Helper to receive from an optional priority mailbox.
-    /// Returns a future that is pending forever if the mailbox is None.
-    async fn recv_priority_opt(
-        priority_mailbox: &mut Option<mpsc::Receiver<StreamerMessage>>,
-    ) -> Option<StreamerMessage> {
-        match priority_mailbox {
-            Some(rx) => rx.recv().await,
-            None => std::future::pending().await,
-        }
-    }
-
-    /// Create a future that completes when the next check is due.
-    ///
-    /// This implements self-scheduling by calculating the delay until
-    /// the next check based on the actor's internal state.
-    ///
-    /// If `duration` is `None`, the timer never fires (waits forever).
-    /// This happens when the streamer is live and no check is scheduled.
-    async fn create_check_timer(duration: Option<Duration>) {
-        match duration {
-            None => {
-                // No check scheduled - wait forever (will be interrupted by other events)
-                std::future::pending::<()>().await;
-            }
-            Some(d) if d.is_zero() => {
-                // Check is due immediately, but yield to allow message processing
-                tokio::task::yield_now().await;
-            }
-            Some(d) => {
-                tokio::time::sleep(d).await;
-            }
-        }
-    }
-
-    fn live_watchdog_interval(&self) -> Duration {
-        // watch dog interval each 2 hours
-        std::cmp::max(
-            Duration::from_secs(2 * 60 * 60),
-            Duration::from_millis(self.config.check_interval_ms),
-        )
-    }
-
-    fn live_stall_watchdog_interval(&self) -> Duration {
-        // If we stop seeing download heartbeats while Live, consider the download "stalled"
-        // and verify status sooner than the 2h watchdog.
-        Duration::from_secs(5 * 60)
-    }
-
-    fn live_watchdog_error_backoff(&self) -> Duration {
-        // Minimum spacing between Live watchdog re-checks after one failed while the
-        // download is unreachable; caps the stall-timer-driven retry cadence.
-        Duration::from_secs(60)
-    }
-
-    fn time_until_live_stall_watchdog(&self) -> Duration {
-        let stall = self.live_stall_watchdog_interval();
-        let Some(last) = self.state.last_download_activity_at else {
-            return stall;
-        };
-
-        let elapsed = last.elapsed();
-        if elapsed >= stall {
-            Duration::ZERO
-        } else {
-            stall - elapsed
-        }
-    }
-
-    /// Whether a timer-driven check right now is the "live watchdog": the
-    /// actor believes the streamer is Live and no check is scheduled (Live
-    /// scheduling parks `next_check`; only the 2h watchdog / stall timers
-    /// wake the loop). Watchdog checks must not mutate DB error/backoff
-    /// state — the download may still be healthy.
-    fn is_live_watchdog(&self) -> bool {
-        self.state.streamer_state == StreamerState::Live && self.state.next_check.is_none()
-    }
-
-    /// Whether no download heartbeat has arrived for the stall window.
-    ///
-    /// While Live, `last_download_activity_at` is fed by
-    /// `handle_download_heartbeat`; a healthy download keeps it fresh. When
-    /// it goes stale the download is gone but no `DownloadEnded` was
-    /// delivered (e.g. the start was dropped by a container gate after the
-    /// session had already resumed).
-    fn download_heartbeats_stale(&self) -> bool {
-        self.state
-            .last_download_activity_at
-            .is_none_or(|t| t.elapsed() >= self.live_stall_watchdog_interval())
-    }
-
-    /// Stall-watchdog reconciliation shared by `perform_check` and
-    /// `handle_batch_result`: a watchdog check that still sees Live while
-    /// download heartbeats have been silent for the stall window means the
-    /// download is dead but the actor was never told. The `(Live, Live)`
-    /// arm of `HysteresisState::should_emit` would suppress the result and
-    /// park this actor forever; resetting the local state makes the check
-    /// register as a NotLive→Live transition so `process_status` re-drives
-    /// the download. Safe when a download is alive but silent: the
-    /// container dedups via `has_active_download` before starting another.
-    ///
-    /// Must run before the caller refreshes `last_download_activity_at`
-    /// from a Live check result — that refresh would mask the staleness.
-    fn force_live_reemit_if_stalled(&mut self, next_state: StreamerState, context: &'static str) {
-        if self.is_live_watchdog()
-            && self.download_heartbeats_stale()
-            && next_state == StreamerState::Live
-        {
-            info!(
-                streamer_id = %self.id,
-                context,
-                "live watchdog: no download heartbeats for the stall window; forcing live re-emit"
-            );
-            self.state.streamer_state = StreamerState::NotLive;
-        }
-    }
-
-    /// Park the actor after a rejected download start: adopt the rejecting
-    /// subsystem's state locally and re-check once its cooldown expires
-    /// (that check re-detects Live and retries the download through the
-    /// normal monitor path). `handle_download_ended` has already cleared
-    /// `last_download_activity_at`.
-    fn schedule_blocked_retry(&mut self, state: StreamerState, retry_after_secs: u64) {
-        self.state.streamer_state = state;
-        self.state
-            .set_next_check(Instant::now().checked_add(Duration::from_secs(retry_after_secs)));
-    }
-
-    fn handle_suppressed_live_status(
-        &mut self,
-        suppression: ProcessStatusSuppression,
-        previous_runtime_state: StreamerActorState,
-    ) {
-        let retry_after = match suppression {
-            ProcessStatusSuppression::Disabled => {
-                debug!(
-                    streamer_id = %self.id,
-                    previous_state = ?previous_runtime_state.streamer_state,
-                    "live status suppressed because streamer is manually disabled"
-                );
-                Duration::from_millis(self.config.check_interval_ms)
-            }
-            ProcessStatusSuppression::TemporarilyDisabled { retry_after } => {
-                let retry_after = retry_after
-                    .unwrap_or_else(|| Duration::from_millis(self.config.check_interval_ms));
-                debug!(
-                    streamer_id = %self.id,
-                    previous_state = ?previous_runtime_state.streamer_state,
-                    retry_after = ?retry_after,
-                    "live status suppressed by temporary disable; reverting actor state"
-                );
-                retry_after
-            }
-        };
-
-        self.state = previous_runtime_state;
-        self.state.last_download_activity_at = None;
-        self.state
-            .set_next_check(Instant::now().checked_add(retry_after));
-    }
-
-    /// Initiate a status check.
-    ///
-    /// If on a batch-capable platform, delegates to the PlatformActor.
-    /// Otherwise, performs the check directly.
-    async fn initiate_check(&mut self) -> Result<(), ActorError> {
-        // Respect temporary backoff (disabled_until) without removing the actor.
-        // This prevents "dead stop" monitoring where the scheduler removes actors
-        // for temporary errors and never respawns them.
-        let metadata = self
-            .get_metadata()
-            .ok_or_else(|| ActorError::fatal("Streamer removed from metadata store"))?;
-        if metadata.is_disabled() {
-            let remaining = metadata.remaining_backoff_std().unwrap_or(Duration::ZERO);
-
-            self.state
-                .set_next_check(Instant::now().checked_add(remaining));
-            debug!(
-                streamer_id = %self.id,
-                streamer_name = %metadata.name,
-                remaining = ?remaining,
-                "status check skipped (backoff)"
-            );
-            return Ok(());
-        }
-
-        debug!(
-            streamer_id = %self.id,
-            streamer_name = %metadata.name,
-            streamer_url = %metadata.url,
-            batch = self.uses_batch_detection(),
-            "status check start"
-        );
-
-        if self.uses_batch_detection() {
-            // Delegate to platform actor for batch detection
-            self.delegate_to_platform().await?;
-        } else {
-            // Perform individual check
-            self.perform_check().await?;
-        }
-
-        Ok(())
-    }
-
-    /// Delegate check to the platform actor for batch processing.
-    async fn delegate_to_platform(&mut self) -> Result<(), ActorError> {
-        if let Some(ref platform_actor) = self.platform_actor {
-            let (reply_tx, reply_rx) = tokio::sync::oneshot::channel();
-
-            let msg = PlatformMessage::RequestCheck {
-                streamer_id: self.id.clone(),
-                reply: reply_tx,
-            };
-
-            // Send request to platform actor
-            if platform_actor.send(msg).await.is_err() {
-                return Err(ActorError::recoverable("Platform actor unavailable"));
-            }
-
-            // Wait for acknowledgment (not the result - that comes via BatchResult message)
-            match tokio::time::timeout(Duration::from_secs(5), reply_rx).await {
-                Ok(Ok(())) => {
-                    // The result arrives asynchronously via handle_batch_result; park
-                    // next_check so the run loop does not re-fire this timer and issue a
-                    // duplicate RequestCheck before that result lands.
-                    self.state
-                        .schedule_next_check(&self.config, self.get_error_count());
-                    debug!("StreamerActor {} check delegated to platform", self.id);
-                    Ok(())
-                }
-                Ok(Err(_)) => Err(ActorError::recoverable("Platform actor dropped reply")),
-                Err(_) => Err(ActorError::recoverable("Platform actor timeout")),
-            }
-        } else {
-            Err(ActorError::recoverable("No platform actor configured"))
-        }
-    }
-
-    /// Perform an individual status check using the configured status checker.
-    ///
-    /// This method connects to the actual monitoring infrastructure via the
-    /// StatusChecker trait, which abstracts the status checking operation.
-    async fn perform_check(&mut self) -> Result<(), ActorError> {
-        // Captured before the check so the result application below cannot
-        // change the answer mid-function; see `is_live_watchdog` for why
-        // watchdog failures must stay side-effect free.
-        let is_live_watchdog = self.is_live_watchdog();
-
-        // Fetch fresh metadata from the store
-        let metadata = self
-            .get_metadata()
-            .ok_or_else(|| ActorError::fatal("Streamer removed from metadata store"))?;
-
-        // Perform the actual status check using the status checker
-        match self.status_checker.check_status(&metadata).await {
-            Ok((result, status)) => {
-                // A successful check clears any watchdog re-check floor set by a prior
-                // failed Live watchdog check.
-                self.live_watchdog_backoff_until = None;
-
-                let previous_runtime_state = self.state.clone();
-                let next_state = result.state;
-                let error_count = self.get_error_count();
-                let reconcile_offline = self.initial_status_pending
-                    && matches!(status, LiveStatus::Offline)
-                    && !self.state.hysteresis.was_live();
-                let settles_initial_status =
-                    matches!(status, LiveStatus::Live { .. } | LiveStatus::Offline);
-
-                self.force_live_reemit_if_stalled(next_state, "check");
-
-                if next_state == StreamerState::Live {
-                    self.state.last_download_activity_at = Some(Instant::now());
-                }
-
-                // Record the check result and get hysteresis decision
-                let should_emit = self.state.record_check(result, &self.config, error_count);
-
-                if should_emit || reconcile_offline {
-                    match self.status_checker.process_status(&metadata, status).await {
-                        Ok(ProcessStatusResult::Applied) => {
-                            if settles_initial_status {
-                                self.initial_status_pending = false;
-                            }
-                        }
-                        Ok(ProcessStatusResult::Suppressed(suppression)) => {
-                            if next_state == StreamerState::Live {
-                                self.handle_suppressed_live_status(
-                                    suppression,
-                                    previous_runtime_state,
-                                );
-                            }
-                        }
-                        Err(e) => {
-                            warn!("StreamerActor {} failed to process status: {}", self.id, e);
-                            // Revert Live state to prevent the actor from getting stuck in
-                            // the watchdog path when no session/download was actually created.
-                            if self.state.streamer_state == StreamerState::Live {
-                                self.state.streamer_state = StreamerState::NotLive;
-                                self.state.last_download_activity_at = None;
-                                self.state.schedule_immediate_check();
-                            }
-                        }
-                    }
-                }
-
-                // Check for fatal states - actor should stop monitoring
-                // Fatal states indicate permanent issues that require manual intervention
-                if next_state == StreamerState::NotFound || next_state == StreamerState::FatalError
-                {
-                    info!(
-                        "StreamerActor {} detected fatal state {:?}, stopping",
-                        self.id, next_state
-                    );
-                    return Err(ActorError::fatal(format!(
-                        "Streamer entered fatal state: {:?}",
-                        next_state
-                    )));
-                }
-
-                debug!(
-                    streamer_id = %self.id,
-                    streamer_name = %metadata.name,
-                    state = ?self.state.streamer_state,
-                    next_check_in = ?self.state.time_until_next_check(),
-                    "status check complete"
-                );
-
-                Ok(())
-            }
-            Err(e) => {
-                if is_live_watchdog {
-                    // Do not call status_checker.handle_error() and do not record an Error state:
-                    // this would increment consecutive error counts, potentially set disabled_until,
-                    // and would also switch scheduling away from the Live watchdog cadence.
-                    // Arm a re-check floor so the Live scheduling branch of `run` does not
-                    // busy-retry against the stall timer while the download is unreachable.
-                    self.live_watchdog_backoff_until =
-                        Some(Instant::now() + self.live_watchdog_error_backoff());
-                    return Err(ActorError::recoverable(format!(
-                        "Live watchdog status check failed (ignored while download active): {}",
-                        e.message
-                    )));
-                }
-
-                // Handle the error through the status checker
-                if let Err(handle_err) = self
-                    .status_checker
-                    .handle_error(&metadata, &e.message)
-                    .await
-                {
-                    warn!(
-                        "StreamerActor {} failed to handle error: {}",
-                        self.id, handle_err
-                    );
-                }
-
-                // Record the error in state
-                let error_result = CheckResult::failure(&e.message);
-                self.state
-                    .record_check(error_result, &self.config, self.get_error_count());
-
-                if e.transient {
-                    // Transient errors are recoverable
-                    Err(ActorError::recoverable(e.message))
-                } else {
-                    // Permanent errors are fatal
-                    Err(ActorError::fatal(e.message))
-                }
-            }
-        }
-    }
-
-    /// Handle an incoming message.
-    ///
-    /// Returns `true` if the actor should stop.
-    async fn handle_message(&mut self, msg: StreamerMessage) -> Result<bool, ActorError> {
-        match msg {
-            StreamerMessage::LifecycleFeedback(envelope) => {
-                use crate::scheduler::feedback::FeedbackDisposition;
-                let result = if self.cancellation_token.is_cancelled() {
-                    Ok(FeedbackDisposition::Retired)
-                } else if envelope.sequence <= self.feedback_sequence
-                    || envelope
-                        .current_epoch
-                        .load(std::sync::atomic::Ordering::Acquire)
-                        != envelope.epoch
-                {
-                    Ok(FeedbackDisposition::Superseded)
-                } else {
-                    self.apply_lifecycle_feedback(
-                        envelope.sequence,
-                        envelope.event.as_ref().clone(),
-                    )
-                    .await
-                };
-                let reply = match &result {
-                    Ok(disposition) => Ok(*disposition),
-                    Err(error) if !error.recoverable => Ok(FeedbackDisposition::Retired),
-                    Err(error) => Err(error.to_string()),
-                };
-                drop(envelope.lease);
-                let _ = envelope.applied.send(reply);
-                result?;
-                Ok(false)
-            }
-            StreamerMessage::RetainedConfig { config, applied } => {
-                if config.revision > self.config_revision {
-                    self.handle_config_update(config.config).await?;
-                    self.config_revision = config.revision;
-                }
-                let _ = applied.send(());
-                Ok(false)
-            }
-            StreamerMessage::RetryAdmission => {
-                if self.current_download.is_none() {
-                    self.state.streamer_state = StreamerState::NotLive;
-                    self.state.schedule_immediate_check();
-                }
-                Ok(false)
-            }
-            StreamerMessage::CheckStatus => {
-                self.handle_check_status().await?;
-                Ok(false)
-            }
-            StreamerMessage::ConfigUpdate(config) => {
-                self.handle_config_update(config).await?;
-                Ok(false)
-            }
-            StreamerMessage::BatchResult(result) => {
-                self.handle_batch_result(*result).await?;
-                Ok(false)
-            }
-            StreamerMessage::DownloadStarted {
-                download_id,
-                session_id,
-            } => {
-                self.handle_download_started(download_id, session_id)
-                    .await?;
-                Ok(false)
-            }
-            StreamerMessage::DownloadHeartbeat {
-                download_id,
-                session_id,
-                progress,
-            } => {
-                self.handle_download_heartbeat(download_id, session_id, progress);
-                Ok(false)
-            }
-            StreamerMessage::DownloadEnded(reason) => {
-                self.current_download = None;
-                self.handle_download_ended(reason).await?;
-                Ok(false)
-            }
-            StreamerMessage::Stop => {
-                self.handle_stop().await?;
-                Ok(true)
-            }
-            StreamerMessage::GetState(reply) => {
-                self.handle_get_state(reply).await;
-                Ok(false)
-            }
-        }
-    }
-
-    /// Handle CheckStatus message - trigger an immediate check.
-    async fn handle_check_status(&mut self) -> Result<(), ActorError> {
-        debug!("StreamerActor {} received CheckStatus", self.id);
-
-        // Reset next check to now to trigger immediate check
-        self.state.set_next_check(Some(Instant::now()));
-
-        Ok(())
-    }
-
-    /// Handle ConfigUpdate message - apply new configuration without restart.
-    ///
-    /// Changed cadence can bring recurring polling forward, but does not postpone
-    /// existing work or advance an admission/backoff/smart-wake deadline.
-    async fn handle_config_update(&mut self, config: StreamerConfig) -> Result<(), ActorError> {
-        debug!("StreamerActor {} received ConfigUpdate", self.id);
-
-        let old_config = std::mem::replace(&mut self.config, config);
-
-        // Log significant changes
-        if old_config.check_interval_ms != self.config.check_interval_ms {
-            info!(
-                "StreamerActor {} check interval changed: {}ms -> {}ms",
-                self.id, old_config.check_interval_ms, self.config.check_interval_ms
-            );
-        }
-
-        if old_config.priority != self.config.priority {
-            info!(
-                "StreamerActor {} priority changed: {:?} -> {:?}",
-                self.id, old_config.priority, self.config.priority
-            );
-        }
-
-        self.state
-            .reschedule_for_config(&self.config, self.get_error_count());
-
-        Ok(())
-    }
-
-    /// Handle BatchResult message - process result from PlatformActor.
-    async fn handle_batch_result(
-        &mut self,
-        result: BatchDetectionResult,
-    ) -> Result<(), ActorError> {
-        debug!(
-            "StreamerActor {} received BatchResult: {:?}",
-            self.id, result.result.state
-        );
-
-        // Verify this result is for us
-        if result.streamer_id != self.id {
-            warn!(
-                "StreamerActor {} received BatchResult for wrong streamer: {}",
-                self.id, result.streamer_id
-            );
-            return Ok(());
-        }
-
-        let previous_runtime_state = self.state.clone();
-        let next_state = result.result.state;
-        let error_count = self.get_error_count();
-        let is_error = result.result.is_error();
-        let error_message = result.result.error.clone();
-
-        // Batch failures are handled as errors, not as offline transitions.
-        // This matches perform_check() behavior and avoids incorrectly ending sessions.
-        if is_error {
-            let is_live_watchdog = self.is_live_watchdog();
-            if is_live_watchdog {
-                let msg = error_message.as_deref().unwrap_or("Batch check failed");
-                warn!(
-                    "StreamerActor {} live watchdog batch check failed (ignored while download active): {}",
-                    self.id, msg
-                );
-                return Ok(());
-            }
-
-            // Record the check result so scheduling/backoff can proceed normally when not Live.
-            self.state
-                .record_check(result.result, &self.config, error_count);
-
-            if let Some(metadata) = self.get_metadata() {
-                let msg = error_message.as_deref().unwrap_or("Batch check failed");
-                if let Err(e) = self.status_checker.handle_error(&metadata, msg).await {
-                    warn!(
-                        "StreamerActor {} failed to handle batch error: {}",
-                        self.id, e
-                    );
-                }
-            }
-            return Ok(());
-        }
-
-        // Record the check result and get hysteresis decision.
-        // Reconciliation must precede the Live-result refresh below.
-        let reconcile_offline = self.initial_status_pending
-            && matches!(result.status, LiveStatus::Offline)
-            && !self.state.hysteresis.was_live();
-        let settles_initial_status =
-            matches!(result.status, LiveStatus::Live { .. } | LiveStatus::Offline);
-        self.force_live_reemit_if_stalled(next_state, "batch");
-
-        if next_state == StreamerState::Live {
-            self.state.last_download_activity_at = Some(Instant::now());
-        }
-
-        let should_emit = self
-            .state
-            .record_check(result.result, &self.config, error_count);
-
-        if should_emit || reconcile_offline {
-            // Fetch fresh metadata for process_status
-            if let Some(metadata) = self.get_metadata() {
-                match self
-                    .status_checker
-                    .process_status(&metadata, result.status)
-                    .await
-                {
-                    Ok(ProcessStatusResult::Applied) => {
-                        if settles_initial_status {
-                            self.initial_status_pending = false;
-                        }
-                    }
-                    Ok(ProcessStatusResult::Suppressed(suppression)) => {
-                        if next_state == StreamerState::Live {
-                            self.handle_suppressed_live_status(suppression, previous_runtime_state);
-                        }
-                    }
-                    Err(e) => {
-                        warn!(
-                            "StreamerActor {} failed to process batch status: {}",
-                            self.id, e
-                        );
-                        // Revert Live state to prevent the actor from getting stuck in
-                        // the watchdog path when no session/download was actually created.
-                        if self.state.streamer_state == StreamerState::Live {
-                            self.state.streamer_state = StreamerState::NotLive;
-                            self.state.last_download_activity_at = None;
-                            self.state.schedule_immediate_check();
-                        }
-                    }
-                }
-            }
-        }
-
-        debug!(
-            "StreamerActor {} batch result processed, next check in {:?}",
-            self.id,
-            self.state.time_until_next_check()
-        );
-
-        Ok(())
-    }
-
-    async fn apply_lifecycle_feedback(
-        &mut self,
-        sequence: u64,
-        event: crate::downloader::DownloadManagerEvent,
-    ) -> Result<crate::scheduler::feedback::FeedbackDisposition, ActorError> {
-        use crate::downloader::{
-            DownloadManagerEvent, DownloadProgressEvent, DownloadTerminalEvent,
-        };
-        use crate::scheduler::feedback::FeedbackDisposition;
-        match event {
-            DownloadManagerEvent::Progress(DownloadProgressEvent::DownloadStarted {
-                download_id,
-                session_id,
-                ..
-            }) => {
-                self.current_download = Some((download_id.clone(), session_id.clone()));
-                self.handle_download_started(download_id, session_id)
-                    .await?;
-            }
-            DownloadManagerEvent::Terminal(terminal) => {
-                if self
-                    .current_download
-                    .as_ref()
-                    .is_some_and(|(download, session)| {
-                        terminal.download_id() != Some(download.as_str())
-                            || terminal.session_id() != session
-                    })
-                {
-                    self.feedback_sequence = sequence;
-                    return Ok(FeedbackDisposition::Superseded);
-                }
-                let policy = match terminal {
-                    DownloadTerminalEvent::Completed { stop_cause, .. } => stop_cause
-                        .map(crate::scheduler::service::download_end_policy_for_stop)
-                        .unwrap_or(DownloadEndPolicy::Completed),
-                    DownloadTerminalEvent::Cancelled { cause, .. } => {
-                        crate::scheduler::service::download_end_policy_for_stop(cause)
-                    }
-                    DownloadTerminalEvent::Failed { error, .. } => {
-                        DownloadEndPolicy::SegmentFailed(error)
-                    }
-                    DownloadTerminalEvent::Rejected {
-                        reason,
-                        retry_after_secs,
-                        session_id,
-                        kind,
-                        ..
-                    } => {
-                        let retry_after_secs = retry_after_secs.unwrap_or(60);
-                        match kind {
-                            crate::downloader::DownloadRejectedKind::CircuitBreaker => {
-                                DownloadEndPolicy::CircuitBreakerBlocked {
-                                    reason,
-                                    retry_after_secs,
-                                    session_id,
-                                }
-                            }
-                            crate::downloader::DownloadRejectedKind::OutputRootUnavailable {
-                                path,
-                                io_kind,
-                            } => DownloadEndPolicy::OutputRootBlocked {
-                                path,
-                                io_kind,
-                                retry_after_secs,
-                                session_id,
-                            },
-                            crate::downloader::DownloadRejectedKind::StreamerBackoff => {
-                                DownloadEndPolicy::StreamerBackoffBlocked {
-                                    reason,
-                                    retry_after_secs,
-                                    session_id,
-                                }
-                            }
-                        }
-                    }
-                };
-                self.current_download = None;
-                self.handle_download_ended(policy).await?;
-            }
-            _ => return Ok(FeedbackDisposition::Superseded),
-        }
-        self.feedback_sequence = sequence;
-        Ok(FeedbackDisposition::Applied)
-    }
-
-    /// Handle DownloadStarted message - pause status checking while a download is active.
-    async fn handle_download_started(
-        &mut self,
-        download_id: String,
-        session_id: String,
-    ) -> Result<(), ActorError> {
-        info!(
-            "StreamerActor {} download started: download_id={}, session_id={}",
-            self.id, download_id, session_id
-        );
-
-        self.seed_active_download(download_id, session_id);
-
-        Ok(())
-    }
-
-    /// Initialize a replacement from the recording owner's current identity.
-    /// This is a snapshot, not a replay of completed session side effects.
-    pub(super) fn seed_active_download(&mut self, download_id: String, session_id: String) {
-        self.current_download = Some((download_id, session_id));
-
-        // Pause checks by switching to Live scheduling behavior.
-        // This is primarily for externally orchestrated downloads where the actor
-        // might not have just observed a Live check result.
-        self.state.streamer_state = StreamerState::Live;
-        self.state.hysteresis.mark_live();
-        self.state.last_download_activity_at = Some(Instant::now());
-        self.state
-            .schedule_next_check(&self.config, self.get_error_count());
-    }
-
-    fn handle_download_heartbeat(
-        &mut self,
-        download_id: String,
-        session_id: String,
-        progress: Option<crate::downloader::engine::DownloadProgress>,
-    ) {
-        if self
-            .current_download
-            .as_ref()
-            .is_none_or(|current| current.0 != download_id || current.1 != session_id)
-        {
-            return;
-        }
-        // Heartbeats are intentionally lightweight; they are used only to avoid triggering
-        // platform extraction while a download is actively making progress.
-        self.state.last_download_activity_at = Some(Instant::now());
-        if let Some(progress) = progress {
-            trace!(
-                "StreamerActor {} download heartbeat: download_id={}, session_id={}, bytes={}, segments={}, speed={}",
-                self.id,
-                download_id,
-                session_id,
-                progress.bytes_downloaded,
-                progress.segments_completed,
-                progress.speed_bytes_per_sec
-            );
-        } else {
-            trace!(
-                "StreamerActor {} download heartbeat: download_id={}, session_id={}",
-                self.id, download_id, session_id
-            );
-        }
-    }
-
-    /// Handle DownloadEnded message - resume status checking or stop monitoring.
-    ///
-    /// # Download End Reason Behaviors
-    ///
-    /// This method handles different download end scenarios with distinct behaviors:
-    ///
-    /// | Reason | Actor Behavior | Hysteresis | Next Check | Rationale |
-    /// |--------|---------------|------------|------------|-----------|
-    /// | `StreamerOffline` | Continue monitoring | Preserve (start post-live short polling) | Offline interval | Stream ended naturally, keep watching for next stream (and catch quick restarts) |
-    /// | `NetworkError` | Continue monitoring | Preserve | Immediate | Technical issue, verify status quickly to resume if still live |
-    /// | `SegmentFailed` | Continue monitoring | Preserve | Immediate | Technical issue, verify status quickly to resume if still live |
-    /// | `UserCancelled` | **STOP ACTOR** | N/A | N/A | User explicitly wants to stop monitoring this streamer |
-    /// | `OutOfSchedule` | Continue monitoring | Preserve | Smart wake / normal | Policy stop; streamer may still be live, but recording is not allowed |
-    /// | `Other` | Continue monitoring | Preserve | Normal interval | Unknown reason, verify actual state through checks |
-    ///
-    /// ## StreamerOffline
-    /// The download orchestration reported the streamer went offline. We immediately publish
-    /// an Offline status to the monitor (to end the session + update DB), then keep the
-    /// hysteresis "recently live" flag so the actor uses the shorter offline polling interval
-    /// for a few checks to catch quick restarts.
-    ///
-    /// ## NetworkError / SegmentFailed
-    /// Technical failures that don't confirm the streamer is offline. We preserve hysteresis
-    /// state and check immediately to quickly resume if the streamer is still live. If they're
-    /// truly offline, the grace period will confirm it through multiple checks.
-    ///
-    /// ## UserCancelled (User-Initiated Stop)
-    ///
-    /// The User stop cause submits offline status. It stops the actor only
-    /// when metadata is cancelled or inactive. The orchestration layer must
-    /// persist `CANCELLED` before delivery; otherwise the actor keeps polling.
-    /// Disable/delete instead ends the session through the runtime coordinator
-    /// before removing the actor.
-    ///
-    /// ## Other
-    /// Unknown/unexpected reasons. We preserve hysteresis and use normal scheduling,
-    /// allowing the grace period to confirm the actual state through status checks.
-    async fn handle_download_ended(&mut self, reason: DownloadEndPolicy) -> Result<(), ActorError> {
-        info!("StreamerActor {} download ended: {:?}", self.id, reason);
-
-        let error_count = self.get_error_count();
-        self.state.last_download_activity_at = None;
-
-        // Update state and schedule check based on reason
-        match reason {
-            DownloadEndPolicy::Stopped(DownloadStopCause::DanmuStreamClosed) => {
-                // The danmu observer owns `handle_offline_with_session` for
-                // this terminal event and records the definitive offline cause.
-                // The actor must not call `process_status(Offline)` again;
-                // it updates only its local scheduling state.
-                self.state.streamer_state = StreamerState::NotLive;
-                self.state.hysteresis.reset();
-                self.state.schedule_next_check(&self.config, error_count);
-            }
-            DownloadEndPolicy::OutOfSchedule
-            | DownloadEndPolicy::Stopped(DownloadStopCause::OutOfSchedule) => {
-                // Policy stop: the streamer may still be live, but the recording window ended.
-                // Do NOT publish Offline; the monitor already recorded OutOfSchedule.
-                //
-                // Keep the actor in OutOfSchedule state so schedule_next_check can use the
-                // smart-wake hint (derived from FilterReason::OutOfSchedule).
-                self.state.streamer_state = StreamerState::OutOfSchedule;
-                if !self.state.hysteresis.was_live() {
-                    // Be robust to externally orchestrated downloads where we never observed Live.
-                    self.state.hysteresis.mark_live();
-                }
-                self.state.schedule_next_check(&self.config, error_count);
-            }
-            DownloadEndPolicy::Completed => {
-                // `on_download_terminal` owns lifecycle updates: clean EOF
-                // enters hysteresis, while HlsEndlist ends authoritatively.
-                // Calling `process_status(Offline)` here would override that
-                // decision. Resume local short polling so the next platform
-                // check can provide an authoritative status.
-                self.state.streamer_state = StreamerState::NotLive;
-                if !self.state.hysteresis.was_live() {
-                    self.state.hysteresis.mark_live();
-                }
-                self.state.hysteresis.mark_offline_observed();
-                self.state.schedule_next_check(&self.config, error_count);
-            }
-            DownloadEndPolicy::Stopped(
-                DownloadStopCause::Shutdown | DownloadStopCause::StreamerDisabled,
-            ) => {
-                // Lifecycle orchestration owns these stops: shutdown preserves the session
-                // for recovery, while disable/delete performs its own authoritative closure.
-                // Park local polling until cancellation/configuration, without publishing Offline.
-                self.state.streamer_state = StreamerState::NotLive;
-                self.state.set_next_check(None);
-            }
-            DownloadEndPolicy::StreamerOffline
-            | DownloadEndPolicy::Stopped(DownloadStopCause::StreamerOffline) => {
-                // Streamer went offline normally. Push an Offline status to the monitor
-                // immediately so DB/session state is updated without waiting for the next check.
-                let metadata = self
-                    .get_metadata()
-                    .ok_or_else(|| ActorError::fatal("Streamer removed from metadata store"))?;
-                if let Err(e) = self
-                    .status_checker
-                    .process_status(&metadata, LiveStatus::Offline)
-                    .await
-                {
-                    warn!(
-                        "StreamerActor {} failed to process offline status on download end: {}",
-                        self.id, e
-                    );
-                }
-
-                // Resume checks using the post-live short polling window.
-                self.state.streamer_state = StreamerState::NotLive;
-                if !self.state.hysteresis.was_live() {
-                    // Be robust to externally orchestrated downloads where we never observed Live.
-                    self.state.hysteresis.mark_live();
-                }
-                self.state.hysteresis.mark_offline_observed();
-
-                // Schedule next check using offline interval while "recently live".
-                self.state.schedule_next_check(&self.config, error_count);
-            }
-            DownloadEndPolicy::NetworkError(_) | DownloadEndPolicy::SegmentFailed(_) => {
-                // Network issue - we don't know if streamer is still live
-                // Schedule immediate check to verify status and potentially resume quickly
-                // Don't reset hysteresis - let the check result determine it
-                self.state.streamer_state = StreamerState::NotLive;
-                self.state.schedule_immediate_check();
-            }
-            DownloadEndPolicy::UserCancelled
-            | DownloadEndPolicy::Stopped(DownloadStopCause::User) => {
-                // User cancelled - stop monitoring this streamer entirely
-                // User intent: "I don't want to monitor this streamer anymore"
-                // The download orchestration layer should update the streamer state
-                // to CANCELLED in the database before sending this message
-
-                // END SESSION BEFORE STOPPING - ensure session is properly closed
-                // This mirrors the StreamerOffline case to maintain consistency
-                let metadata = self
-                    .get_metadata()
-                    .ok_or_else(|| ActorError::fatal("Streamer removed from metadata store"))?;
-                if let Err(e) = self
-                    .status_checker
-                    .process_status(&metadata, LiveStatus::Offline)
-                    .await
-                {
-                    warn!(
-                        "StreamerActor {} failed to process offline status on cancellation: {}",
-                        self.id, e
-                    );
-                }
-
-                if metadata.state == StreamerState::Cancelled || !metadata.is_active() {
-                    info!(
-                        "StreamerActor {} stopping due to user cancellation",
-                        self.id
-                    );
-                    return Err(ActorError::fatal(
-                        "User cancelled download - stopping monitoring",
-                    ));
-                }
-
-                // Orchestration bug / inconsistent state: avoid stopping the actor if the
-                // streamer is still active, otherwise the supervisor will restart it.
-                warn!(
-                    "StreamerActor {} received UserCancelled but streamer state is still {:?}; continuing monitoring",
-                    self.id, metadata.state
-                );
-                self.state.streamer_state = StreamerState::NotLive;
-                if !self.state.hysteresis.was_live() {
-                    // Be robust to externally orchestrated downloads where we never observed Live.
-                    self.state.hysteresis.mark_live();
-                }
-                self.state.hysteresis.mark_offline_observed();
-                self.state.schedule_next_check(&self.config, error_count);
-            }
-            DownloadEndPolicy::Other(_)
-            | DownloadEndPolicy::Stopped(DownloadStopCause::Other(_)) => {
-                // Unknown reason - don't reset hysteresis, let status checks verify state
-                // If streamer is truly offline, grace period will confirm it through multiple checks
-                self.state.streamer_state = StreamerState::NotLive;
-                self.state.schedule_next_check(&self.config, error_count);
-            }
-            DownloadEndPolicy::CircuitBreakerBlocked {
-                reason,
-                retry_after_secs,
-                ..
-            } => {
-                // Circuit breaker blocked the download - this is a temporal error.
-                // Set state to TemporalDisabled for visibility (user can see something is wrong)
-                // and schedule check after cooldown.
-                info!(
-                    "StreamerActor {} download blocked by circuit breaker: {}, retry in {}s",
-                    self.id, reason, retry_after_secs
-                );
-
-                // Persist to DB so UI shows correct state
-                let metadata = self
-                    .get_metadata()
-                    .ok_or_else(|| ActorError::fatal("Streamer removed from metadata store"))?;
-                if let Err(e) = self
-                    .status_checker
-                    .set_infra_blocked(
-                        &metadata,
-                        crate::monitor::InfraBlockReason::CircuitBreaker { retry_after_secs },
-                    )
-                    .await
-                {
-                    warn!(
-                        "StreamerActor {} failed to persist infra block state: {}",
-                        self.id, e
-                    );
-                }
-
-                self.schedule_blocked_retry(StreamerState::TemporalDisabled, retry_after_secs);
-            }
-            DownloadEndPolicy::OutputRootBlocked {
-                path,
-                io_kind,
-                retry_after_secs,
-                ..
-            } => {
-                // Output-root write gate blocked the download — the recording
-                // filesystem is unwritable. This is an infrastructure-level
-                // issue (not a per-streamer fault), so we transition to
-                // OutOfSpace via set_infra_blocked (which does NOT bump
-                // consecutive_error_count) and schedule the next check after
-                // the gate cooldown. When the gate recovers, its
-                // recovery_hook will clear this streamer's disabled_until and
-                // last_error so it re-enters the live-check rotation
-                // immediately on the next tick.
-                info!(
-                    "StreamerActor {} download blocked by output root gate: {} ({:?}), retry in {}s",
-                    self.id,
-                    path.display(),
-                    io_kind,
-                    retry_after_secs
-                );
-
-                let metadata = self
-                    .get_metadata()
-                    .ok_or_else(|| ActorError::fatal("Streamer removed from metadata store"))?;
-                if let Err(e) = self
-                    .status_checker
-                    .set_infra_blocked(
-                        &metadata,
-                        crate::monitor::InfraBlockReason::OutputRootUnavailable {
-                            path: path.clone(),
-                            io_kind,
-                            retry_after_secs,
-                        },
-                    )
-                    .await
-                {
-                    warn!(
-                        "StreamerActor {} failed to persist output-root block state: {}",
-                        self.id, e
-                    );
-                }
-
-                self.schedule_blocked_retry(StreamerState::OutOfSpace, retry_after_secs);
-            }
-            DownloadEndPolicy::StreamerBackoffBlocked {
-                reason,
-                retry_after_secs,
-                ..
-            } => {
-                // The download-start gate refused to start because the
-                // streamer is inside its `disabled_until` error backoff. The
-                // backoff was already persisted by
-                // `MonitorService::handle_error`, so unlike the
-                // CircuitBreakerBlocked arm there is no `set_infra_blocked`
-                // write to make. Only local scheduling needs correcting: the
-                // live check that raced the backoff commit left this actor
-                // in Live with no download, where the `(Live, Live)` arm of
-                // `HysteresisState::should_emit` would keep `process_status`
-                // from ever firing again.
-                info!(
-                    streamer_id = %self.id,
-                    reason = %reason,
-                    retry_after_secs,
-                    "download blocked by streamer error backoff; rescheduling check at expiry"
-                );
-
-                self.schedule_blocked_retry(StreamerState::TemporalDisabled, retry_after_secs);
-            }
-        }
-
-        debug!(
-            "StreamerActor {} resuming status checks, next check in {:?}",
-            self.id,
-            self.state.time_until_next_check()
-        );
-
-        Ok(())
-    }
-
-    /// Handle Stop message - prepare for graceful shutdown.
-    async fn handle_stop(&mut self) -> Result<(), ActorError> {
-        info!("StreamerActor {} received Stop", self.id);
-
-        Ok(())
-    }
-
-    /// Handle GetState message - return current state via oneshot channel.
-    async fn handle_get_state(&self, reply: tokio::sync::oneshot::Sender<StreamerActorState>) {
-        debug!("StreamerActor {} received GetState", self.id);
-
-        // Send state, ignore if receiver dropped
-        let _ = reply.send(self.state.clone());
-    }
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn config_messages_preserve_recurring_and_authoritative_retry_deadlines() {
-        let config = create_test_config();
-        let (mut actor, _) = StreamerActor::new(
-            "test-streamer".to_owned(),
-            create_test_metadata_store(),
-            config.clone(),
-            CancellationToken::new(),
-            create_noop_checker(),
-        );
-        actor.state.schedule_next_check(&config, 0);
-        let recurring = actor.state.next_check;
-        actor
-            .handle_config_update(StreamerConfig {
-                priority: Priority::High,
-                ..config.clone()
-            })
-            .await
-            .unwrap();
-        assert_eq!(actor.state.next_check, recurring);
-        actor.schedule_blocked_retry(StreamerState::TemporalDisabled, 30);
-        let retry = actor.state.next_check;
-        actor
-            .handle_config_update(StreamerConfig {
-                check_interval_ms: 1,
-                ..config
-            })
-            .await
-            .unwrap();
-        assert_eq!(
-            actor.state.next_check, retry,
-            "a shorter configured interval cannot advance an admission cooldown"
-        );
-    }
-    use crate::domain::Priority;
-    use crate::monitor::{ProcessStatusResult, ProcessStatusSuppression};
-    use crate::scheduler::actor::monitor_adapter::{CheckError, NoOpStatusChecker};
-    use async_trait::async_trait;
-    use std::collections::VecDeque;
-    use std::sync::Arc;
-    use std::sync::Mutex;
-
-    fn create_test_metadata() -> StreamerMetadata {
-        StreamerMetadata {
-            id: "test-streamer".to_string(),
-            name: "Test Streamer".to_string(),
-            url: "https://twitch.tv/test".to_string(),
-            platform_config_id: "twitch".to_string(),
-            template_config_id: None,
-            state: StreamerState::NotLive,
-            priority: Priority::Normal,
-            avatar_url: None,
-            consecutive_error_count: 0,
-            disabled_until: None,
-            last_live_time: None,
-            last_error: None,
-            streamer_specific_config: None,
-            offline_check_count: 3,
-            offline_check_delay_ms: 20_000,
-            created_at: chrono::Utc::now(),
-            deleted_at: None,
-            updated_at: chrono::Utc::now(),
-        }
-    }
-
-    fn create_test_metadata_store() -> Arc<DashMap<String, Arc<StreamerMetadata>>> {
-        let store = Arc::new(DashMap::new());
-        let metadata = create_test_metadata();
-        store.insert(metadata.id.clone(), Arc::new(metadata));
-        store
-    }
-
-    fn create_test_config() -> StreamerConfig {
-        StreamerConfig {
-            check_interval_ms: 1000, // 1 second for tests
-            offline_check_interval_ms: 500,
-            offline_check_count: 3,
-            priority: Priority::Normal,
-            batch_capable: false,
-        }
-    }
-
-    fn create_noop_checker() -> Arc<dyn StatusChecker> {
-        Arc::new(NoOpStatusChecker)
-    }
-
-    #[derive(Debug)]
-    struct SequenceStatusChecker {
-        checks: Mutex<VecDeque<(CheckResult, LiveStatus)>>,
-        outcomes: Mutex<VecDeque<Result<ProcessStatusResult, CheckError>>>,
-    }
-
-    impl SequenceStatusChecker {
-        fn new(checks: Vec<(CheckResult, LiveStatus)>, outcomes: Vec<ProcessStatusResult>) -> Self {
-            Self {
-                checks: Mutex::new(VecDeque::from(checks)),
-                outcomes: Mutex::new(outcomes.into_iter().map(Ok).collect()),
-            }
-        }
-    }
-
-    #[async_trait]
-    impl StatusChecker for SequenceStatusChecker {
-        async fn check_status(
-            &self,
-            _streamer: &StreamerMetadata,
-        ) -> Result<(CheckResult, LiveStatus), CheckError> {
-            self.checks
-                .lock()
-                .unwrap()
-                .pop_front()
-                .ok_or_else(|| CheckError::transient("missing check result"))
-        }
-
-        async fn process_status(
-            &self,
-            _streamer: &StreamerMetadata,
-            _status: LiveStatus,
-        ) -> Result<ProcessStatusResult, CheckError> {
-            self.outcomes
-                .lock()
-                .unwrap()
-                .pop_front()
-                .unwrap_or(Ok(ProcessStatusResult::Applied))
-        }
-
-        async fn handle_error(
-            &self,
-            _streamer: &StreamerMetadata,
-            _error: &str,
-        ) -> Result<(), CheckError> {
-            Ok(())
-        }
-
-        async fn set_infra_blocked(
-            &self,
-            _streamer: &StreamerMetadata,
-            _reason: crate::monitor::InfraBlockReason,
-        ) -> Result<(), CheckError> {
-            Ok(())
-        }
-    }
-
-    #[tokio::test]
-    async fn initial_offline_reconciliation_retries_until_applied_then_suppresses_duplicates() {
-        for batch in [false, true] {
-            let checker = Arc::new(SequenceStatusChecker::new(
-                (0..4)
-                    .map(|_| {
-                        (
-                            CheckResult::success(StreamerState::NotLive),
-                            LiveStatus::Offline,
-                        )
-                    })
-                    .collect(),
-                vec![
-                    ProcessStatusResult::Suppressed(
-                        ProcessStatusSuppression::TemporarilyDisabled {
-                            retry_after: Some(Duration::from_secs(1)),
-                        },
-                    ),
-                    ProcessStatusResult::Applied,
-                    ProcessStatusResult::Applied,
-                ],
-            ));
-            checker
-                .outcomes
-                .lock()
-                .unwrap()
-                .push_front(Err(CheckError::transient("database busy")));
-            let (mut actor, _handle) = if batch {
-                StreamerActor::with_priority_channel(
-                    "test-streamer".to_string(),
-                    create_test_metadata_store(),
-                    create_test_config(),
-                    CancellationToken::new(),
-                    checker.clone(),
-                )
-            } else {
-                StreamerActor::new(
-                    "test-streamer".to_string(),
-                    create_test_metadata_store(),
-                    create_test_config(),
-                    CancellationToken::new(),
-                    checker.clone(),
-                )
-            };
-            for (pending, remaining) in [(true, 3), (true, 2), (false, 1), (false, 1)] {
-                if batch {
-                    actor
-                        .handle_batch_result(BatchDetectionResult {
-                            streamer_id: "test-streamer".to_string(),
-                            result: CheckResult::success(StreamerState::NotLive),
-                            status: LiveStatus::Offline,
-                        })
-                        .await
-                        .unwrap();
-                } else {
-                    actor.perform_check().await.unwrap();
-                }
-                assert_eq!(actor.initial_status_pending, pending);
-                assert_eq!(checker.outcomes.lock().unwrap().len(), remaining);
-                assert!(actor.state.next_check.is_some());
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn live_actor_retries_failed_offline_at_grace_threshold() {
-        for batch in [false, true] {
-            let mut previous_metadata = create_test_metadata();
-            previous_metadata.state = StreamerState::Live;
-            let previous_state = StreamerActorState::from_metadata(&previous_metadata);
-            let checker = Arc::new(SequenceStatusChecker::new(
-                (0..5)
-                    .map(|_| {
-                        (
-                            CheckResult::success(StreamerState::NotLive),
-                            LiveStatus::Offline,
-                        )
-                    })
-                    .collect(),
-                vec![ProcessStatusResult::Applied, ProcessStatusResult::Applied],
-            ));
-            checker
-                .outcomes
-                .lock()
-                .unwrap()
-                .push_front(Err(CheckError::transient("database busy")));
-            let (mut actor, _handle) = if batch {
-                StreamerActor::with_priority_channel(
-                    "test-streamer".to_string(),
-                    create_test_metadata_store(),
-                    create_test_config(),
-                    CancellationToken::new(),
-                    checker.clone(),
-                )
-            } else {
-                StreamerActor::new(
-                    "test-streamer".to_string(),
-                    create_test_metadata_store(),
-                    create_test_config(),
-                    CancellationToken::new(),
-                    checker.clone(),
-                )
-            };
-            actor.state = previous_state;
-            assert_eq!(actor.state.streamer_state, StreamerState::Live);
-            for (pending, remaining) in [(true, 3), (true, 3), (true, 2), (false, 1), (false, 1)] {
-                if batch {
-                    actor
-                        .handle_batch_result(BatchDetectionResult {
-                            streamer_id: "test-streamer".to_string(),
-                            result: CheckResult::success(StreamerState::NotLive),
-                            status: LiveStatus::Offline,
-                        })
-                        .await
-                        .unwrap();
-                } else {
-                    actor.perform_check().await.unwrap();
-                }
-                assert_eq!(actor.initial_status_pending, pending);
-                assert_eq!(checker.outcomes.lock().unwrap().len(), remaining);
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn initial_reconciliation_keeps_known_live_offline_grace_period() {
-        for batch in [false, true] {
-            let store = create_test_metadata_store();
-            Arc::make_mut(&mut store.get_mut("test-streamer").unwrap()).state = StreamerState::Live;
-            let checker = Arc::new(SequenceStatusChecker::new(
-                (0..3)
-                    .map(|_| {
-                        (
-                            CheckResult::success(StreamerState::NotLive),
-                            LiveStatus::Offline,
-                        )
-                    })
-                    .collect(),
-                vec![ProcessStatusResult::Applied],
-            ));
-            let (mut actor, _handle) = StreamerActor::new(
-                "test-streamer".to_string(),
-                store,
-                create_test_config(),
-                CancellationToken::new(),
-                checker.clone(),
-            );
-            for remaining in [1, 1, 0] {
-                if batch {
-                    actor
-                        .handle_batch_result(BatchDetectionResult {
-                            streamer_id: "test-streamer".to_string(),
-                            result: CheckResult::success(StreamerState::NotLive),
-                            status: LiveStatus::Offline,
-                        })
-                        .await
-                        .unwrap();
-                } else {
-                    actor.perform_check().await.unwrap();
-                }
-                assert_eq!(checker.outcomes.lock().unwrap().len(), remaining);
-            }
-            assert!(!actor.initial_status_pending);
-        }
-    }
-
-    #[test]
-    fn test_streamer_actor_new() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (actor, handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            create_noop_checker(),
-        );
-
-        assert_eq!(actor.id(), "test-streamer");
-        assert_eq!(handle.id(), "test-streamer");
-        assert!(!actor.uses_batch_detection());
-    }
-
-    #[test]
-    fn test_streamer_actor_with_platform() {
-        let metadata_store = create_test_metadata_store();
-        let mut config = create_test_config();
-        config.batch_capable = true;
-        let token = CancellationToken::new();
-        let (platform_tx, _platform_rx) = mpsc::channel::<PlatformMessage>(10);
-
-        let (actor, _handle) = StreamerActor::with_priority_and_platform(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            platform_tx,
-            create_noop_checker(),
-        );
-
-        assert!(actor.uses_batch_detection());
-    }
-
-    #[tokio::test]
-    async fn test_streamer_actor_get_state() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (actor, handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token.clone(),
-            create_noop_checker(),
-        );
-
-        // Spawn actor
-        let actor_task = tokio::spawn(async move { actor.run().await });
-
-        // Query state
-        let (reply_tx, reply_rx) = tokio::sync::oneshot::channel();
-        handle
-            .send(StreamerMessage::GetState(reply_tx))
-            .await
-            .unwrap();
-
-        let state = reply_rx.await.unwrap();
-        assert_eq!(state.streamer_state, StreamerState::NotLive);
-        assert_eq!(state.hysteresis.offline_count(), 0);
-
-        // Stop actor
-        handle.send(StreamerMessage::Stop).await.unwrap();
-        let result = actor_task.await.unwrap();
-        assert!(matches!(result, Ok(ActorOutcome::Stopped)));
-    }
-
-    #[tokio::test]
-    async fn test_streamer_actor_config_update() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (actor, handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token.clone(),
-            create_noop_checker(),
-        );
-
-        // Spawn actor
-        let actor_task = tokio::spawn(async move { actor.run().await });
-
-        // Send config update
-        let new_config = StreamerConfig {
-            check_interval_ms: 5000,
-            offline_check_interval_ms: 2000,
-            offline_check_count: 5,
-            priority: Priority::High,
-            batch_capable: false,
-        };
-        handle
-            .send(StreamerMessage::ConfigUpdate(new_config))
-            .await
-            .unwrap();
-
-        // Give time for processing
-        tokio::time::sleep(Duration::from_millis(50)).await;
-
-        // Query state to verify config was applied
-        let (reply_tx, reply_rx) = tokio::sync::oneshot::channel();
-        handle
-            .send(StreamerMessage::GetState(reply_tx))
-            .await
-            .unwrap();
-        let _state = reply_rx.await.unwrap();
-
-        // Stop actor
-        handle.send(StreamerMessage::Stop).await.unwrap();
-        let result = actor_task.await.unwrap();
-        assert!(matches!(result, Ok(ActorOutcome::Stopped)));
-    }
-
-    #[tokio::test]
-    async fn test_streamer_actor_cancellation() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token.clone(),
-            create_noop_checker(),
-        );
-
-        // Spawn actor
-        let actor_task = tokio::spawn(async move { actor.run().await });
-
-        // Cancel
-        token.cancel();
-
-        let result = actor_task.await.unwrap();
-        assert!(matches!(result, Ok(ActorOutcome::Cancelled)));
-    }
-
-    #[tokio::test]
-    async fn test_streamer_actor_batch_result() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (actor, handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token.clone(),
-            create_noop_checker(),
-        );
-
-        // Spawn actor
-        let actor_task = tokio::spawn(async move { actor.run().await });
-
-        // Send batch result
-        let batch_result = BatchDetectionResult {
-            streamer_id: "test-streamer".to_string(),
-            result: CheckResult::success(StreamerState::Live),
-            status: crate::monitor::LiveStatus::Live {
-                title: "Test Stream".to_string(),
-                category: None,
-                started_at: None,
-                viewer_count: None,
-                avatar: None,
-                streams: vec![],
-                media_headers: None,
-                media_extras: None,
-                next_check_hint: None,
-                candidates: vec![],
-            },
-        };
-        handle
-            .send(StreamerMessage::BatchResult(Box::new(batch_result)))
-            .await
-            .unwrap();
-
-        // Give time for processing
-        tokio::time::sleep(Duration::from_millis(50)).await;
-
-        // Query state
-        let (reply_tx, reply_rx) = tokio::sync::oneshot::channel();
-        handle
-            .send(StreamerMessage::GetState(reply_tx))
-            .await
-            .unwrap();
-        let state = reply_rx.await.unwrap();
-
-        // State should be updated to Live
-        assert_eq!(state.streamer_state, StreamerState::Live);
-
-        // Stop actor
-        handle.send(StreamerMessage::Stop).await.unwrap();
-        let result = actor_task.await.unwrap();
-        assert!(matches!(result, Ok(ActorOutcome::Stopped)));
-    }
-
-    #[test]
-    fn test_actor_error_display() {
-        let err = ActorError::recoverable("test error");
-        assert_eq!(err.to_string(), "test error");
-        assert!(err.recoverable);
-
-        let err = ActorError::fatal("fatal error");
-        assert_eq!(err.to_string(), "fatal error");
-        assert!(!err.recoverable);
-    }
-
-    #[test]
-    fn test_actor_outcome() {
-        assert_eq!(ActorOutcome::Stopped, ActorOutcome::Stopped);
-        assert_ne!(ActorOutcome::Stopped, ActorOutcome::Cancelled);
-    }
-
-    #[test]
-    fn test_streamer_actor_with_priority_channel() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (actor, handle) = StreamerActor::with_priority_channel(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            create_noop_checker(),
-        );
-
-        assert_eq!(actor.id(), "test-streamer");
-        assert_eq!(handle.id(), "test-streamer");
-        // Actor should have priority mailbox
-        assert!(actor.priority_mailbox.is_some());
-    }
-
-    #[test]
-    fn test_streamer_actor_with_priority_and_platform() {
-        let metadata_store = create_test_metadata_store();
-        let mut config = create_test_config();
-        config.batch_capable = true;
-        let token = CancellationToken::new();
-        let (platform_tx, _platform_rx) = mpsc::channel::<PlatformMessage>(10);
-
-        let (actor, _handle) = StreamerActor::with_priority_and_platform(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            platform_tx,
-            create_noop_checker(),
-        );
-
-        assert!(actor.uses_batch_detection());
-        assert!(actor.priority_mailbox.is_some());
-    }
-
-    #[tokio::test]
-    async fn test_priority_channel_stop_message() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (actor, handle) = StreamerActor::with_priority_channel(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token.clone(),
-            create_noop_checker(),
-        );
-
-        // Spawn actor
-        let actor_task = tokio::spawn(async move { actor.run().await });
-
-        // Send stop via priority channel
-        handle.send_priority(StreamerMessage::Stop).await.unwrap();
-
-        let result = actor_task.await.unwrap();
-        assert!(matches!(result, Ok(ActorOutcome::Stopped)));
-    }
-
-    #[tokio::test]
-    async fn test_priority_channel_processes_before_normal() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (actor, handle) = StreamerActor::with_priority_channel(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token.clone(),
-            create_noop_checker(),
-        );
-
-        // Spawn actor
-        let actor_task = tokio::spawn(async move { actor.run().await });
-
-        // Send multiple normal messages first
-        for _ in 0..5 {
-            let batch_result = BatchDetectionResult {
-                streamer_id: "test-streamer".to_string(),
-                result: CheckResult::success(StreamerState::NotLive),
-                status: crate::monitor::LiveStatus::Offline,
-            };
-            handle
-                .send(StreamerMessage::BatchResult(Box::new(batch_result)))
-                .await
-                .unwrap();
-        }
-
-        // Send stop via priority channel - should be processed promptly
-        handle.send_priority(StreamerMessage::Stop).await.unwrap();
-
-        // Actor should stop quickly despite pending normal messages
-        let result = tokio::time::timeout(Duration::from_millis(500), actor_task)
-            .await
-            .unwrap()
-            .unwrap();
-
-        assert!(matches!(result, Ok(ActorOutcome::Stopped)));
-    }
-
-    #[tokio::test]
-    async fn test_priority_channel_get_state() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (actor, handle) = StreamerActor::with_priority_channel(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token.clone(),
-            create_noop_checker(),
-        );
-
-        // Spawn actor
-        let actor_task = tokio::spawn(async move { actor.run().await });
-
-        // Query state via priority channel
-        let (reply_tx, reply_rx) = tokio::sync::oneshot::channel();
-        handle
-            .send_priority(StreamerMessage::GetState(reply_tx))
-            .await
-            .unwrap();
-
-        let state = reply_rx.await.unwrap();
-        assert_eq!(state.streamer_state, StreamerState::NotLive);
-
-        // Stop actor
-        handle.send_priority(StreamerMessage::Stop).await.unwrap();
-        let result = actor_task.await.unwrap();
-        assert!(matches!(result, Ok(ActorOutcome::Stopped)));
-    }
-
-    #[tokio::test]
-    async fn test_streamer_actor_resume_on_download_end() {
-        let metadata_store = create_test_metadata_store();
-        let metadata_store_for_update = metadata_store.clone();
-        let config = StreamerConfig::default();
-        let token = CancellationToken::new();
-
-        // Create actor with live state (checks paused)
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config.clone(),
-            token,
-            create_noop_checker(),
-        );
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live(); // Set was_live since we were live
-        actor.state.next_check = None;
-
-        // Verify initially paused
-        assert!(actor.state.next_check.is_none());
-
-        // Simulate download ended (streamer offline)
-        let result = actor
-            .handle_download_ended(super::super::messages::DownloadEndPolicy::StreamerOffline)
-            .await;
-        assert!(result.is_ok());
-
-        // Verify state changed and check scheduled
-        assert_eq!(actor.state.streamer_state, StreamerState::NotLive);
-        assert!(actor.state.next_check.is_some());
-        // Hysteresis stays "recently live" so short offline polling is used
-        assert!(actor.state.hysteresis.was_live());
-        assert_eq!(actor.state.hysteresis.offline_count(), 1);
-        let until = actor.state.time_until_next_check().unwrap();
-        assert_eq!(
-            actor.state.recurring_interval_ms,
-            Some(config.offline_check_interval_ms)
-        );
-        assert!(until <= Duration::from_millis(config.offline_check_interval_ms * 11 / 10));
-
-        // Reset and test error case
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live();
-        actor.state.next_check = None;
-
-        // Simulate download failed (network error)
-        let result = actor
-            .handle_download_ended(super::super::messages::DownloadEndPolicy::NetworkError(
-                "timeout".into(),
-            ))
-            .await;
-        assert!(result.is_ok());
-
-        // Verify state changed (assumed not live for safety) and check scheduled immediately
-        assert_eq!(actor.state.streamer_state, StreamerState::NotLive);
-        assert!(actor.state.next_check.is_some());
-        assert!(actor.state.is_check_due()); // Network errors trigger immediate check
-
-        // Reset and test user cancellation
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live();
-        actor.state.next_check = None;
-
-        // Simulate user cancelled download
-        if let Some(mut entry) = metadata_store_for_update.get_mut("test-streamer") {
-            Arc::make_mut(&mut entry).state = StreamerState::Cancelled;
-        }
-        let result = actor
-            .handle_download_ended(super::super::messages::DownloadEndPolicy::UserCancelled)
-            .await;
-
-        // Should return fatal error since user wants to stop monitoring
-        assert!(result.is_err());
-        let err = result.unwrap_err();
-        assert!(!err.recoverable); // Fatal error stops the actor
-        assert!(err.message.contains("User cancelled"));
-
-        // Reset and test unknown reason
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live();
-        actor.state.next_check = None;
-
-        // Simulate download ended with unknown reason
-        let result = actor
-            .handle_download_ended(super::super::messages::DownloadEndPolicy::Other(
-                "unknown".into(),
-            ))
-            .await;
-        assert!(result.is_ok());
-
-        // Verify state changed but hysteresis preserved (let checks verify)
-        assert_eq!(actor.state.streamer_state, StreamerState::NotLive);
-        assert!(actor.state.next_check.is_some());
-        // Hysteresis should be preserved - let checks determine actual state
-        assert!(actor.state.hysteresis.was_live());
-    }
-
-    /// `DownloadEndPolicy::Completed` (engine clean EOF without platform
-    /// authority) must update local scheduling state to the post-live
-    /// short-polling cadence — same as `StreamerOffline` — but must NOT
-    /// emit a `process_status(Offline)` push (the lifecycle's hysteresis
-    /// is the authoritative absorber for this case; a duplicate offline
-    /// emit would race and bypass it).
-    ///
-    /// We can't directly assert "no push" here without a tracking
-    /// `StatusChecker`, but we can lock in the local-state contract:
-    /// the post-live cadence assertions identical to
-    /// `test_streamer_actor_resume_on_download_end` for the
-    /// `StreamerOffline` path. If a future change re-routes `Completed`
-    /// through a different scheduling branch, this test fails and points
-    /// the maintainer at the contract.
-    #[tokio::test]
-    async fn test_streamer_actor_completed_uses_post_live_cadence() {
-        let metadata_store = create_test_metadata_store();
-        let config = StreamerConfig::default();
-        let token = CancellationToken::new();
-
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config.clone(),
-            token,
-            create_noop_checker(),
-        );
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live();
-        actor.state.next_check = None;
-
-        let result = actor
-            .handle_download_ended(super::super::messages::DownloadEndPolicy::Completed)
-            .await;
-        assert!(result.is_ok());
-
-        assert_eq!(actor.state.streamer_state, StreamerState::NotLive);
-        assert!(actor.state.next_check.is_some());
-        // Same post-live cadence as `StreamerOffline`: was_live preserved,
-        // offline-observed counter incremented, next check arrives within
-        // the jittered offline polling window.
-        assert!(actor.state.hysteresis.was_live());
-        assert_eq!(actor.state.hysteresis.offline_count(), 1);
-        let until = actor.state.time_until_next_check().unwrap();
-        assert!(
-            until <= Duration::from_millis(config.offline_check_interval_ms * 11 / 10),
-            "Completed must schedule next check within the post-live offline window"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_perform_check_suppressed_live_does_not_leave_actor_stuck_live() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let checker: Arc<dyn StatusChecker> = Arc::new(SequenceStatusChecker::new(
-            vec![(
-                CheckResult::success(StreamerState::Live),
-                LiveStatus::Live {
-                    title: "Suppressed Live".to_string(),
-                    category: None,
-                    started_at: None,
-                    viewer_count: None,
-                    avatar: None,
-                    streams: vec![],
-                    media_headers: None,
-                    media_extras: None,
-                    next_check_hint: None,
-                    candidates: vec![],
-                },
-            )],
-            vec![ProcessStatusResult::Suppressed(
-                ProcessStatusSuppression::TemporarilyDisabled {
-                    retry_after: Some(Duration::from_secs(30)),
-                },
-            )],
-        ));
-
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            checker,
-        );
-
-        actor.perform_check().await.unwrap();
-
-        assert_eq!(actor.state.streamer_state, StreamerState::NotLive);
-        assert!(actor.state.last_download_activity_at.is_none());
-        assert!(actor.state.next_check.is_some());
-        assert!(!actor.state.hysteresis.was_live());
-
-        let until = actor.state.time_until_next_check().unwrap();
-        assert!(until <= Duration::from_secs(30));
-    }
-
-    #[tokio::test]
-    async fn test_perform_check_recovers_after_suppressed_live_when_backoff_expires() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let checker: Arc<dyn StatusChecker> = Arc::new(SequenceStatusChecker::new(
-            vec![
-                (
-                    CheckResult::success(StreamerState::Live),
-                    LiveStatus::Live {
-                        title: "Suppressed Live".to_string(),
-                        category: None,
-                        started_at: None,
-                        viewer_count: None,
-                        avatar: None,
-                        streams: vec![],
-                        media_headers: None,
-                        media_extras: None,
-                        next_check_hint: None,
-                        candidates: vec![],
-                    },
-                ),
-                (
-                    CheckResult::success(StreamerState::Live),
-                    LiveStatus::Live {
-                        title: "Recovered Live".to_string(),
-                        category: None,
-                        started_at: None,
-                        viewer_count: None,
-                        avatar: None,
-                        streams: vec![],
-                        media_headers: None,
-                        media_extras: None,
-                        next_check_hint: None,
-                        candidates: vec![],
-                    },
-                ),
-            ],
-            vec![
-                ProcessStatusResult::Suppressed(ProcessStatusSuppression::TemporarilyDisabled {
-                    retry_after: Some(Duration::from_secs(1)),
-                }),
-                ProcessStatusResult::Applied,
-            ],
-        ));
-
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            checker,
-        );
-
-        actor.perform_check().await.unwrap();
-        assert_eq!(actor.state.streamer_state, StreamerState::NotLive);
-        assert!(!actor.state.hysteresis.was_live());
-
-        actor.perform_check().await.unwrap();
-
-        assert_eq!(actor.state.streamer_state, StreamerState::Live);
-        assert!(actor.state.last_download_activity_at.is_some());
-        assert!(actor.state.next_check.is_none());
-        assert!(actor.state.hysteresis.was_live());
-    }
-
-    /// A live-watchdog check that still sees Live while download heartbeats
-    /// have been silent for the stall window must NOT be swallowed by the
-    /// `(Live, Live)` hysteresis suppression — it must call `process_status`
-    /// so the monitor can re-drive the download (a dead session otherwise
-    /// stays unrecovered: no download → no DownloadEnded → no transition).
-    #[tokio::test]
-    async fn test_live_watchdog_stall_forces_live_reemit() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let checker = Arc::new(SequenceStatusChecker::new(
-            vec![(
-                CheckResult::success(StreamerState::Live),
-                LiveStatus::Live {
-                    title: "Stalled Live".to_string(),
-                    category: None,
-                    started_at: None,
-                    viewer_count: None,
-                    avatar: None,
-                    streams: vec![],
-                    media_headers: None,
-                    media_extras: None,
-                    next_check_hint: None,
-                    candidates: vec![],
-                },
-            )],
-            vec![ProcessStatusResult::Applied],
-        ));
-
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            checker.clone() as Arc<dyn StatusChecker>,
-        );
-
-        // Wedged shape: Live with no scheduled check (watchdog mode) and no
-        // download heartbeat ever observed.
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live();
-        actor.state.next_check = None;
-        actor.state.last_download_activity_at = None;
-
-        actor.perform_check().await.unwrap();
-
-        // process_status consumed its queued outcome — the Live result was
-        // re-emitted instead of suppressed.
-        assert!(
-            checker.outcomes.lock().unwrap().is_empty(),
-            "stalled live watchdog must call process_status"
-        );
-        assert_eq!(actor.state.streamer_state, StreamerState::Live);
-        assert!(actor.state.last_download_activity_at.is_some());
-        assert!(actor.state.hysteresis.was_live());
-    }
-
-    /// The reconciliation must not fire while a download is healthy: fresh
-    /// heartbeats mean the `(Live, Live)` suppression is doing its intended
-    /// job (avoiding redundant monitor writes on the 2h watchdog).
-    #[tokio::test]
-    async fn test_live_watchdog_with_fresh_heartbeats_keeps_suppression() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let checker = Arc::new(SequenceStatusChecker::new(
-            vec![(
-                CheckResult::success(StreamerState::Live),
-                LiveStatus::Live {
-                    title: "Healthy Live".to_string(),
-                    category: None,
-                    started_at: None,
-                    viewer_count: None,
-                    avatar: None,
-                    streams: vec![],
-                    media_headers: None,
-                    media_extras: None,
-                    next_check_hint: None,
-                    candidates: vec![],
-                },
-            )],
-            vec![ProcessStatusResult::Applied],
-        ));
-
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            checker.clone() as Arc<dyn StatusChecker>,
-        );
-
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live();
-        actor.state.next_check = None;
-        // Heartbeat just arrived — download is alive.
-        actor.state.last_download_activity_at = Some(Instant::now());
-
-        actor.perform_check().await.unwrap();
-
-        assert_eq!(
-            checker.outcomes.lock().unwrap().len(),
-            1,
-            "healthy live watchdog must keep the (Live, Live) suppression"
-        );
-        assert_eq!(actor.state.streamer_state, StreamerState::Live);
-        assert!(actor.state.next_check.is_none());
-    }
-
-    /// `StreamerBackoffBlocked` (container refused the download start because
-    /// the streamer is inside its `disabled_until` error backoff) must move
-    /// the actor out of Live and schedule a re-check at backoff expiry —
-    /// otherwise the actor sits in watchdog mode where `(Live, Live)`
-    /// suppression never lets `process_status` restart the download.
-    #[tokio::test]
-    async fn test_download_ended_streamer_backoff_blocked_schedules_expiry_check() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            create_noop_checker(),
-        );
-
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live();
-        actor.state.next_check = None;
-        actor.state.last_download_activity_at = Some(Instant::now());
-
-        actor
-            .handle_download_ended(
-                super::super::messages::DownloadEndPolicy::StreamerBackoffBlocked {
-                    reason: "streamer temporarily disabled (error backoff)".to_string(),
-                    retry_after_secs: 30,
-                    session_id: "session-1".to_string(),
-                },
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(actor.state.streamer_state, StreamerState::TemporalDisabled);
-        assert!(actor.state.last_download_activity_at.is_none());
-        let until = actor.state.time_until_next_check().unwrap();
-        assert!(until <= Duration::from_secs(30));
-        assert!(until > Duration::from_secs(20));
-        // TemporalDisabled → Live is an emitting transition in
-        // `HysteresisState::should_emit`, so the expiry check re-drives the
-        // download through the normal monitor path.
-    }
-
-    #[tokio::test]
-    async fn test_suppressed_live_restores_notlive_grace_hysteresis_context() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let checker: Arc<dyn StatusChecker> = Arc::new(SequenceStatusChecker::new(
-            vec![(
-                CheckResult::success(StreamerState::Live),
-                LiveStatus::Live {
-                    title: "Suppressed Live".to_string(),
-                    category: None,
-                    started_at: None,
-                    viewer_count: None,
-                    avatar: None,
-                    streams: vec![],
-                    media_headers: None,
-                    media_extras: None,
-                    next_check_hint: None,
-                    candidates: vec![],
-                },
-            )],
-            vec![ProcessStatusResult::Suppressed(
-                ProcessStatusSuppression::TemporarilyDisabled {
-                    retry_after: Some(Duration::from_secs(30)),
-                },
-            )],
-        ));
-
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config.clone(),
-            token,
-            checker,
-        );
-
-        actor.state.streamer_state = StreamerState::NotLive;
-        actor.state.hysteresis.mark_live();
-        actor.state.hysteresis.mark_offline_observed();
-        let original_offline_count = actor.state.hysteresis.offline_count();
-        actor.state.last_check = Some(CheckResult {
-            state: StreamerState::NotLive,
-            stream_url: None,
-            title: Some("Previous offline".to_string()),
-            checked_at: chrono::Utc::now(),
-            error: None,
-            next_check_hint: None,
-        });
-
-        actor.perform_check().await.unwrap();
-
-        assert_eq!(actor.state.streamer_state, StreamerState::NotLive);
-        assert!(actor.state.hysteresis.was_live());
-        assert_eq!(
-            actor.state.hysteresis.offline_count(),
-            original_offline_count
-        );
-        assert_eq!(
-            actor
-                .state
-                .last_check
-                .as_ref()
-                .and_then(|check| check.title.as_deref()),
-            Some("Previous offline")
-        );
-        let until = actor.state.time_until_next_check().unwrap();
-        assert!(until <= Duration::from_secs(30));
-    }
-
-    #[tokio::test]
-    async fn test_suppressed_live_restores_out_of_schedule_smart_wake_context() {
-        let metadata_store = create_test_metadata_store();
-        let config = create_test_config();
-        let token = CancellationToken::new();
-
-        let checker: Arc<dyn StatusChecker> = Arc::new(SequenceStatusChecker::new(
-            vec![(
-                CheckResult::success(StreamerState::Live),
-                LiveStatus::Live {
-                    title: "Suppressed Live".to_string(),
-                    category: None,
-                    started_at: None,
-                    viewer_count: None,
-                    avatar: None,
-                    streams: vec![],
-                    media_headers: None,
-                    media_extras: None,
-                    next_check_hint: None,
-                    candidates: vec![],
-                },
-            )],
-            vec![ProcessStatusResult::Suppressed(
-                ProcessStatusSuppression::TemporarilyDisabled {
-                    retry_after: Some(Duration::from_secs(30)),
-                },
-            )],
-        ));
-
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            checker,
-        );
-
-        let smart_wake_hint = chrono::Utc::now() + chrono::Duration::minutes(15);
-        actor.state.streamer_state = StreamerState::OutOfSchedule;
-        actor.state.hysteresis.mark_live();
-        actor.state.last_check = Some(CheckResult {
-            state: StreamerState::OutOfSchedule,
-            stream_url: None,
-            title: Some("Out of schedule".to_string()),
-            checked_at: chrono::Utc::now(),
-            error: None,
-            next_check_hint: Some(smart_wake_hint),
-        });
-
-        actor.perform_check().await.unwrap();
-
-        assert_eq!(actor.state.streamer_state, StreamerState::OutOfSchedule);
-        assert!(actor.state.hysteresis.was_live());
-        assert_eq!(
-            actor
-                .state
-                .last_check
-                .as_ref()
-                .and_then(|check| check.next_check_hint),
-            Some(smart_wake_hint)
-        );
-        let until = actor.state.time_until_next_check().unwrap();
-        assert!(until <= Duration::from_secs(30));
-    }
-
-    /// `process_status` must NOT be called on the `Stopped(DanmuStreamClosed)`
-    /// path — the container's danmu observer is the canonical owner of the
-    /// offline emission for this signal. Calling `process_status` here too
-    /// would race the observer at the `streamer.state == Live` gate inside
-    /// `handle_offline_with_session`, ending whichever session the actor's
-    /// snapshot saw as active (frequently a different `session_id` than
-    /// the one the observer is closing).
-    #[derive(Debug, Default)]
-    struct AssertNotCalledStatusChecker;
-
-    #[async_trait]
-    impl StatusChecker for AssertNotCalledStatusChecker {
-        async fn check_status(
-            &self,
-            _streamer: &StreamerMetadata,
-        ) -> Result<(CheckResult, LiveStatus), CheckError> {
-            panic!("AssertNotCalledStatusChecker::check_status must not be called");
-        }
-
-        async fn process_status(
-            &self,
-            _streamer: &StreamerMetadata,
-            _status: LiveStatus,
-        ) -> Result<ProcessStatusResult, CheckError> {
-            panic!("AssertNotCalledStatusChecker::process_status must not be called");
-        }
-
-        async fn handle_error(
-            &self,
-            _streamer: &StreamerMetadata,
-            _error: &str,
-        ) -> Result<(), CheckError> {
-            panic!("AssertNotCalledStatusChecker::handle_error must not be called");
-        }
-
-        async fn set_infra_blocked(
-            &self,
-            _streamer: &StreamerMetadata,
-            _reason: crate::monitor::InfraBlockReason,
-        ) -> Result<(), CheckError> {
-            panic!("AssertNotCalledStatusChecker::set_infra_blocked must not be called");
-        }
-    }
-
-    #[tokio::test]
-    async fn danmu_stream_closed_does_not_call_process_status() {
-        let metadata_store = create_test_metadata_store();
-        let config = StreamerConfig::default();
-        let token = CancellationToken::new();
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            Arc::new(AssertNotCalledStatusChecker),
-        );
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live();
-
-        // No panic ⇒ the actor did NOT route through `process_status`.
-        let result = actor
-            .handle_download_ended(super::super::messages::DownloadEndPolicy::Stopped(
-                DownloadStopCause::DanmuStreamClosed,
-            ))
-            .await;
-        assert!(
-            result.is_ok(),
-            "actor must complete cleanly without process_status"
-        );
-    }
-
-    #[tokio::test]
-    async fn danmu_stream_closed_resets_hysteresis_and_reschedules() {
-        let metadata_store = create_test_metadata_store();
-        let config = StreamerConfig::default();
-        let token = CancellationToken::new();
-        let (mut actor, _handle) = StreamerActor::new(
-            "test-streamer".to_string(),
-            metadata_store,
-            config,
-            token,
-            create_noop_checker(),
-        );
-        actor.state.streamer_state = StreamerState::Live;
-        actor.state.hysteresis.mark_live();
-        actor.state.hysteresis.mark_offline_observed();
-        assert!(
-            actor.state.hysteresis.was_live(),
-            "fixture: hysteresis records the prior live state"
-        );
-        actor.state.next_check = None;
-
-        actor
-            .handle_download_ended(super::super::messages::DownloadEndPolicy::Stopped(
-                DownloadStopCause::DanmuStreamClosed,
-            ))
-            .await
-            .unwrap();
-
-        // The arm flips local in-memory state to NotLive, fully resets
-        // hysteresis (was_live becomes false → schedule_next_check uses
-        // the longer normal interval), and arms a next-check timer.
-        assert_eq!(actor.state.streamer_state, StreamerState::NotLive);
-        assert!(
-            !actor.state.hysteresis.was_live(),
-            "DanmuStreamClosed is authoritative; hysteresis must be fully reset"
-        );
-        assert!(actor.state.next_check.is_some());
-    }
-    #[tokio::test]
-    async fn fatal_timer_and_mailbox_errors_stop_gracefully() {
-        for trigger in ["timer", "normal", "priority", "late-priority"] {
-            let metadata = create_test_metadata_store();
-            let (mut actor, handle) = StreamerActor::with_priority_channel(
-                "test-streamer".to_owned(),
-                metadata.clone(),
-                create_test_config(),
-                CancellationToken::new(),
-                Arc::new(AssertNotCalledStatusChecker),
-            );
-            actor.state.next_check = Some(Instant::now() + Duration::from_secs(3600));
-            let message = StreamerMessage::DownloadEnded(DownloadEndPolicy::StreamerOffline);
-            if trigger != "late-priority" {
-                metadata.remove("test-streamer");
-            }
-            match trigger {
-                "timer" => actor.state.next_check = Some(Instant::now()),
-                "normal" => handle.send(message).await.unwrap(),
-                "priority" => handle.send_priority(message).await.unwrap(),
-                _ => {}
-            }
-            let task = tokio::spawn(actor.run());
-            if trigger == "late-priority" {
-                let (reply, received) = tokio::sync::oneshot::channel();
-                handle.send(StreamerMessage::GetState(reply)).await.unwrap();
-                tokio::time::timeout(Duration::from_secs(2), received)
-                    .await
-                    .unwrap()
-                    .unwrap();
-                metadata.remove("test-streamer");
-                handle
-                    .send_priority(StreamerMessage::DownloadEnded(
-                        DownloadEndPolicy::StreamerOffline,
-                    ))
-                    .await
-                    .unwrap();
-            }
-            let outcome = tokio::time::timeout(Duration::from_secs(2), task)
-                .await
-                .unwrap()
-                .unwrap();
-            assert_eq!(outcome.unwrap(), ActorOutcome::Stopped, "{trigger}");
-        }
-    }
-
-    #[tokio::test]
-    async fn orchestration_stops_and_unknown_causes_never_publish_offline() {
-        for cause in [
-            DownloadStopCause::Shutdown,
-            DownloadStopCause::StreamerDisabled,
-            DownloadStopCause::Other("internal stop".to_owned()),
-        ] {
-            let park = matches!(
-                cause,
-                DownloadStopCause::Shutdown | DownloadStopCause::StreamerDisabled
-            );
-            let (mut actor, _handle) = StreamerActor::new(
-                "test-streamer".to_owned(),
-                create_test_metadata_store(),
-                create_test_config(),
-                CancellationToken::new(),
-                Arc::new(AssertNotCalledStatusChecker),
-            );
-            actor.state.streamer_state = StreamerState::Live;
-            actor.state.hysteresis.mark_live();
-            actor.state.last_download_activity_at = Some(Instant::now());
-            actor
-                .handle_download_ended(DownloadEndPolicy::Stopped(cause))
-                .await
-                .unwrap();
-            assert_eq!(actor.state.streamer_state, StreamerState::NotLive);
-            assert!(
-                actor.state.hysteresis.was_live(),
-                "an orchestration stop is not an offline observation"
-            );
-            assert_eq!(actor.state.hysteresis.offline_count(), 0);
-            assert!(actor.state.last_download_activity_at.is_none());
-            assert_eq!(actor.state.next_check.is_none(), park);
-        }
-    }
-
-    #[tokio::test]
-    async fn authoritative_offline_still_emits_exactly_one_monitor_signal() {
-        for reason in [
-            DownloadEndPolicy::StreamerOffline,
-            DownloadEndPolicy::Stopped(DownloadStopCause::StreamerOffline),
-        ] {
-            let checker = Arc::new(SequenceStatusChecker::new(
-                vec![],
-                vec![ProcessStatusResult::Applied, ProcessStatusResult::Applied],
-            ));
-            let (mut actor, _handle) = StreamerActor::new(
-                "test-streamer".to_owned(),
-                create_test_metadata_store(),
-                create_test_config(),
-                CancellationToken::new(),
-                checker.clone(),
-            );
-            actor.state.streamer_state = StreamerState::Live;
-            actor.state.hysteresis.mark_live();
-            actor.handle_download_ended(reason).await.unwrap();
-            assert_eq!(
-                checker.outcomes.lock().unwrap().len(),
-                1,
-                "offline must reach process_status exactly once"
-            );
-            assert_eq!(actor.state.hysteresis.offline_count(), 1);
-            assert!(actor.state.next_check.is_some());
-        }
-    }
-    #[tokio::test]
-    async fn reliable_feedback_maps_all_terminal_and_rejection_categories() {
-        use crate::downloader::engine::EngineType;
-        use crate::downloader::{
-            DownloadFailureKind, DownloadManagerEvent, DownloadProtocol, DownloadRejectedKind,
-            DownloadTerminalEvent, EngineEndSignal,
-        };
-        use crate::scheduler::feedback::FeedbackDisposition;
-        let completed = |cause| DownloadTerminalEvent::Completed {
-            download_id: "d".to_owned(),
-            streamer_id: "test-streamer".to_owned(),
-            streamer_name: "Test".to_owned(),
-            session_id: "s".to_owned(),
-            total_bytes: 1,
-            total_duration_secs: 1.0,
-            total_segments: 1,
-            file_path: None,
-            engine_signal: EngineEndSignal::CleanDisconnect,
-            stop_cause: cause,
-        };
-        let rejected = |kind| DownloadTerminalEvent::Rejected {
-            streamer_id: "test-streamer".to_owned(),
-            streamer_name: "Test".to_owned(),
-            session_id: "s".to_owned(),
-            reason: "fixture".to_owned(),
-            retry_after_secs: Some(60),
-            kind,
-        };
-        let mut cases = vec![
-            (completed(None), StreamerState::NotLive, true),
-            (
-                completed(Some(DownloadStopCause::OutOfSchedule)),
-                StreamerState::OutOfSchedule,
-                true,
-            ),
-            (
-                completed(Some(DownloadStopCause::Shutdown)),
-                StreamerState::NotLive,
-                false,
-            ),
-            (
-                rejected(DownloadRejectedKind::CircuitBreaker),
-                StreamerState::TemporalDisabled,
-                true,
-            ),
-            (
-                rejected(DownloadRejectedKind::OutputRootUnavailable {
-                    path: "/recordings".into(),
-                    io_kind: crate::downloader::IoErrorKindSer::StorageFull,
-                }),
-                StreamerState::OutOfSpace,
-                true,
-            ),
-            (
-                rejected(DownloadRejectedKind::StreamerBackoff),
-                StreamerState::TemporalDisabled,
-                true,
-            ),
-            (
-                DownloadTerminalEvent::Failed {
-                    download_id: "d".to_owned(),
-                    streamer_id: "test-streamer".to_owned(),
-                    streamer_name: "Test".to_owned(),
-                    session_id: "s".to_owned(),
-                    engine_type: EngineType::Ffmpeg,
-                    protocol: DownloadProtocol::Flv,
-                    kind: DownloadFailureKind::Other,
-                    error: "fixture".to_owned(),
-                    recoverable: true,
-                },
-                StreamerState::NotLive,
-                true,
-            ),
-        ];
-        for cause in [
-            DownloadStopCause::Shutdown,
-            DownloadStopCause::StreamerDisabled,
-            DownloadStopCause::StreamerOffline,
-            DownloadStopCause::DanmuStreamClosed,
-            DownloadStopCause::OutOfSchedule,
-            DownloadStopCause::Other("fixture".to_owned()),
-            DownloadStopCause::User,
-        ] {
-            let scheduled = !matches!(
-                cause,
-                DownloadStopCause::Shutdown | DownloadStopCause::StreamerDisabled
-            );
-            let expected = if matches!(cause, DownloadStopCause::OutOfSchedule) {
-                StreamerState::OutOfSchedule
-            } else {
-                StreamerState::NotLive
-            };
-            cases.push((
-                DownloadTerminalEvent::Cancelled {
-                    download_id: "d".to_owned(),
-                    streamer_id: "test-streamer".to_owned(),
-                    streamer_name: "Test".to_owned(),
-                    session_id: "s".to_owned(),
-                    cause,
-                },
-                expected,
-                scheduled,
-            ));
-        }
-        for (terminal, expected, scheduled) in cases {
-            let (mut actor, _handle) = StreamerActor::new(
-                "test-streamer".to_owned(),
-                create_test_metadata_store(),
-                create_test_config(),
-                CancellationToken::new(),
-                create_noop_checker(),
-            );
-            assert_eq!(
-                actor
-                    .apply_lifecycle_feedback(1, DownloadManagerEvent::Terminal(terminal))
-                    .await
-                    .unwrap(),
-                FeedbackDisposition::Applied
-            );
-            assert_eq!(actor.state.streamer_state, expected);
-            assert_eq!(actor.state.next_check.is_some(), scheduled);
-        }
-    }
-}
+mod tests;
